+++ mkv-filter-metadata/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.11] - 2026-06-21
+
+### Added
+- Implemented a 3-way System/Light/Dark UI theme configuration option in the Settings page, allowing users to forcefully override the OS default theme. 
+
+### Changed
+- Promoted application to version 1.1.11.
+
 ## [1.1.10] - 2026-06-20
 
 ### Added